+++ mex/common/models/activity.py
@@ -13,6 +13,7 @@ from mex.common.models.base.rules import (
     PreventiveRule,
     RuleSet,
     SubtractiveRule,
+    WorkflowRule,
 )
 from mex.common.types import (
     ActivityType,
@@ -26,6 +27,7 @@ from mex.common.types import (
     MergedOrganizationIdentifier,
     MergedPersonIdentifier,
     MergedPrimarySourceIdentifier,
+    PublishingTarget,
     Text,
     Theme,
     Year,
@@ -426,12 +428,25 @@ class PreventiveActivity(_Stem, PreventiveRule):
     website: list[MergedPrimarySourceIdentifier] = []
 
 
+class WorkflowActivity(_Stem, WorkflowRule):
+    """Rule to prevent publishing of merged activity items."""
+
+    entityType: Annotated[
+        Literal["WorkflowActivity"], Field(alias="$type", frozen=True)
+    ] = "WorkflowActivity"
+    forbiddenPublishingTarget: Annotated[
+        list[PublishingTarget],
+        Field(description="Targets to which the merged item should NOT be published."),
+    ] = []
+
+
 class _BaseRuleSet(_Stem, RuleSet):
     """Base class for sets of rules for an activity item."""
 
     additive: AdditiveActivity = AdditiveActivity()
     subtractive: SubtractiveActivity = SubtractiveActivity()
     preventive: PreventiveActivity = PreventiveActivity()
+    workflow: WorkflowActivity = WorkflowActivity()
 
 
 class ActivityRuleSetRequest(_BaseRuleSet):