+++ mex/common/models/organization.py
@@ -13,11 +13,13 @@ from mex.common.models.base.rules import (
     PreventiveRule,
     RuleSet,
     SubtractiveRule,
+    WorkflowRule,
 )
 from mex.common.types import (
     ExtractedOrganizationIdentifier,
     MergedOrganizationIdentifier,
     MergedPrimarySourceIdentifier,
+    PublishingTarget,
     Text,
 )
 
@@ -362,12 +364,25 @@ class PreventiveOrganization(_Stem, PreventiveRule):
     wikidataId: list[MergedPrimarySourceIdentifier] = []
 
 
+class WorkflowOrganization(_Stem, WorkflowRule):
+    """Rule to prevent publishing of merged organization items."""
+
+    entityType: Annotated[
+        Literal["WorkflowOrganization"], Field(alias="$type", frozen=True)
+    ] = "WorkflowOrganization"
+    forbiddenPublishingTarget: Annotated[
+        list[PublishingTarget],
+        Field(description="Targets to which the merged item should NOT be published."),
+    ] = []
+
+
 class _BaseRuleSet(_Stem, RuleSet):
     """Base class for sets of rules for an organization item."""
 
     additive: AdditiveOrganization = AdditiveOrganization()
     subtractive: SubtractiveOrganization = SubtractiveOrganization()
     preventive: PreventiveOrganization = PreventiveOrganization()
+    workflow: WorkflowOrganization = WorkflowOrganization()
 
 
 class OrganizationRuleSetRequest(_BaseRuleSet):