+++ mex/common/models/organizational_unit.py
@@ -13,6 +13,7 @@ from mex.common.models.base.rules import (
     PreventiveRule,
     RuleSet,
     SubtractiveRule,
+    WorkflowRule,
 )
 from mex.common.types import (
     ExtractedOrganizationalUnitIdentifier,
@@ -20,6 +21,7 @@ from mex.common.types import (
     MergedOrganizationalUnitIdentifier,
     MergedOrganizationIdentifier,
     MergedPrimarySourceIdentifier,
+    PublishingTarget,
     Text,
 )
 
@@ -321,12 +323,25 @@ class PreventiveOrganizationalUnit(_Stem, PreventiveRule):
     website: list[MergedPrimarySourceIdentifier] = []
 
 
+class WorkflowOrganizationalUnit(_Stem, WorkflowRule):
+    """Rule to prevent publishing of merged organizational unit items."""
+
+    entityType: Annotated[
+        Literal["WorkflowOrganizationalUnit"], Field(alias="$type", frozen=True)
+    ] = "WorkflowOrganizationalUnit"
+    forbiddenPublishingTarget: Annotated[
+        list[PublishingTarget],
+        Field(description="Targets to which the merged item should NOT be published."),
+    ] = []
+
+
 class _BaseRuleSet(_Stem, RuleSet):
     """Base class for sets of rules for an organizational unit item."""
 
     additive: AdditiveOrganizationalUnit = AdditiveOrganizationalUnit()
     subtractive: SubtractiveOrganizationalUnit = SubtractiveOrganizationalUnit()
     preventive: PreventiveOrganizationalUnit = PreventiveOrganizationalUnit()
+    workflow: WorkflowOrganizationalUnit = WorkflowOrganizationalUnit()
 
 
 class OrganizationalUnitRuleSetRequest(_BaseRuleSet):