+++ mex/common/models/resource.py
@@ -13,6 +13,7 @@ from mex.common.models.base.rules import (
     PreventiveRule,
     RuleSet,
     SubtractiveRule,
+    WorkflowRule,
 )
 from mex.common.types import (
     AccessRestriction,
@@ -35,6 +36,7 @@ from mex.common.types import (
     MergedPrimarySourceIdentifier,
     MergedResourceIdentifier,
     PersonalData,
+    PublishingTarget,
     ResourceCreationMethod,
     ResourceTypeGeneral,
     Text,
@@ -925,12 +927,25 @@ class PreventiveResource(_Stem, PreventiveRule):
     wasGeneratedBy: list[MergedPrimarySourceIdentifier] = []
 
 
+class WorkflowResource(_Stem, WorkflowRule):
+    """Rule to prevent publishing of merged resource items."""
+
+    entityType: Annotated[
+        Literal["WorkflowResource"], Field(alias="$type", frozen=True)
+    ] = "WorkflowResource"
+    forbiddenPublishingTarget: Annotated[
+        list[PublishingTarget],
+        Field(description="Targets to which the merged item should NOT be published."),
+    ] = []
+
+
 class _BaseRuleSet(_Stem, RuleSet):
     """Base class for sets of rules for a resource item."""
 
     additive: AdditiveResource = AdditiveResource()
     subtractive: SubtractiveResource = SubtractiveResource()
     preventive: PreventiveResource = PreventiveResource()
+    workflow: WorkflowResource = WorkflowResource()
 
 
 class ResourceRuleSetRequest(_BaseRuleSet):