+++ mex/common/models/variable.py
@@ -13,6 +13,7 @@ from mex.common.models.base.rules import (
     PreventiveRule,
     RuleSet,
     SubtractiveRule,
+    WorkflowRule,
 )
 from mex.common.types import (
     ExtractedVariableIdentifier,
@@ -20,6 +21,7 @@ from mex.common.types import (
     MergedResourceIdentifier,
     MergedVariableGroupIdentifier,
     MergedVariableIdentifier,
+    PublishingTarget,
     Text,
 )
 
@@ -356,12 +358,25 @@ class PreventiveVariable(_Stem, PreventiveRule):
     valueSet: list[MergedPrimarySourceIdentifier] = []
 
 
+class WorkflowVariable(_Stem, WorkflowRule):
+    """Rule to prevent publishing of merged variable items."""
+
+    entityType: Annotated[
+        Literal["WorkflowVariable"], Field(alias="$type", frozen=True)
+    ] = "WorkflowVariable"
+    forbiddenPublishingTarget: Annotated[
+        list[PublishingTarget],
+        Field(description="Targets to which the merged item should NOT be published."),
+    ] = []
+
+
 class _BaseRuleSet(_Stem, RuleSet):
     """Base class for sets of rules for a variable item."""
 
     additive: AdditiveVariable = AdditiveVariable()
     subtractive: SubtractiveVariable = SubtractiveVariable()
     preventive: PreventiveVariable = PreventiveVariable()
+    workflow: WorkflowVariable = WorkflowVariable()
 
 
 class VariableRuleSetRequest(_BaseRuleSet):